+++ lib/server.py
@@ -1,0 +1,12 @@
+from flask import Flask, render_template
+
+
+server = Flask(__name__)
+# app = dash.Dash(__name__, server=server,url_base_pathname='/')
+
+jwt = 'Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6ImtleS1pZC0xIiwidHlwIjoiSldUIn0.eyJqdGkiOiI5YmNjYjEzNzAzODM0MzIxYjMxYWY1ZWFkNTEzOTc5NiIsInN1YiI6InRlY2h1c3IiLCJzY29wZSI6WyJtZHNwOmNvcmU6QWRtaW4zcmRQYXJ0eVRlY2hVc2VyIl0sImNsaWVudF9pZCI6InRlY2h1c3IiLCJjaWQiOiJ0ZWNodXNyIiwiYXpwIjoidGVjaHVzciIsImdyYW50X3R5cGUiOiJjbGllbnRfY3JlZGVudGlhbHMiLCJyZXZfc2lnIjoiZmYwMjU3NTAiLCJpYXQiOjE1Mjk1OTA0NDksImV4cCI6MTUyOTU5MjI0OSwiaXNzIjoiaHR0cHM6Ly9zaWxkZXZtcy5waWFtLmV1MS5taW5kc3BoZXJlLmlvL29hdXRoL3Rva2VuIiwiemlkIjoic2lsZGV2bXMiLCJhdWQiOlsidGVjaHVzciJdLCJ0ZW4iOiJzaWxkZXZtcyIsInNjaGVtYXMiOlsidXJuOnNpZW1lbnM6bWluZHNwaGVyZTppYW06djEiXSwiY2F0IjoiY2xpZW50LXRva2VuOnYxIn0.a2CDi3ge9vliDYie-H1Pw8XJVJ-mqwS6lkr3eaGh5YEfGZkbSeqTPaZ7l9ks7DUji4-nesbO4hmn9d3-owSA6izgwRXuelOmsXX5ErHI0YgmHlWR8-Z16wAsrmXW_HvrpRXIM2Qhdj2xQFrzP0S_u9Io9Qu-5mF5UKsQcB8qHjCvM65SkCbC50d2szSWsT84NZRv0ruHv1V89Y-r9ibKCdpKoQjTi6uH3gJGykW_cO4hm_q_x3tsCxe1ViRPq3z7iQ9DkSnM6V9uQ6GBmYD8_watajX2f2nbEtfXD32Y3LI8wfNkx5wVWixu_W2MYwi7jpYPN0ABrBeiCy_xVM4xxg'
+
+@server.route('/')
+def root():
+    return render_template("home.html")
+
