+++ requirements.txt
@@ -1,3 +1,6 @@
-flask
-requests
-numpy
+Flask==0.12.2
+numpy==1.14.2
+dash==0.21.1
+requests==2.18.4
+dash_html_components==0.11.0
+dash_core_components==0.23.0
