+++ templates/includes/_navbar.html
@@ -19,6 +19,12 @@
                   Assets
                 </a>
               </li>
+              <li class="nav-item mr-auto">
+                <a class="nav-link" href="/floorplan">
+                  <span data-feather="map"></span>
+                  Floorplan
+                </a>
+              </li>
             </ul>
           </div>
         </nav>
